market-maven backend/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
finnhub-python==2.4.18
Flask==2.3.2
Flask-Cors==4.0.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multitasking==0.0.11
numpy==1.24.4
packaging==23.1
pandas==2.0.3
plotly==5.15.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
yfinance==0.2.25
zipp==3.16.2
textblob==0.9.0
vaderSentiment==3.3.2
cohere==4.17.0
transformers==4.31.0